server/requirements.txt
Flask==2.0.1
Pillow==8.3.2
psycopg2-binary==2.9.1
